Add UUID to Item and implement item assignment functionality
The code now adds a UUID to each Item at the point of creation for unique identification. In addition, the store's ability to assign a quantity of a unit of an Item has been implemented and tested. Necessary updates have been made to the 'assignment' package and associated tests, and the 'uptake' property has been introduced to the assignment.